requirements.txt
Django
Pillow
psycopg2-binary
djangorestframework==3.11.0
python-decouple
drf-yasg
